+++ src/utils/extractParameterNames.ts
@@ -1,6 +1,4 @@
-import type { string_name } from '../types/typeAliases';
-import type { string_parameter_name } from '../types/typeAliases';
-import type { string_template } from '../types/typeAliases';
+import type { string_name, string_parameter_name, string_template } from '../types/typeAliases';
 
 /**
  * Parses the template and returns the list of all parameter names
@@ -8,7 +6,7 @@ import type { string_template } from '../types/typeAliases';
  * @param template the template with parameters in {curly} braces
  * @returns the list of parameter names
  */
-export function extractParameters(template: string_template): Set<string_parameter_name> {
+export function extractParameterNames(template: string_template): Set<string_parameter_name> {
     const matches = template.matchAll(/{\w+}/g);
     const parameterNames = new Set<string_name>();
     for (const match of matches) {
@@ -19,7 +17,6 @@ export function extractParameters(template: string_template): Set<string_paramet
 
     return parameterNames;
 }
-
 
 /**
  * TODO: !!!!! Rename to extractParameterNames